This must merge after part I so the tests will pass. Fixes definition file. Moves method from local_bgs to claimant_web_service. Creates test file for new service.
