+++ modules/claims_api/app/clients/claims_api/bgs_client/definitions.rb
@@ -61,11 +61,11 @@ module ClaimsApi
           )
       end
 
-      module ClaimantService
+      module ClaimantWebService
         DEFINITION =
           Service.new(
             bean: ClaimantServiceBean::DEFINITION,
-            path: 'ClaimantService'
+            path: 'ClaimantWebService'
           )
       end
