+++ modules/claims_api/lib/bgs_service/claimant_web_service.rb
@@ -25,5 +25,18 @@ module ClaimsApi
       response = make_request(endpoint: bean_name, action: 'addFlash', body:)
       response&.dig(:body, :add_flash_response) || response
     end
+
+    def find_poa_by_participant_id(id)
+      body = Nokogiri::XML::DocumentFragment.parse <<~EOXML
+        <ptcpntId />
+      EOXML
+
+      { ptcpntId: id }.each do |k, v|
+        body.xpath("./*[local-name()='#{k}']")[0].content = v
+      end
+
+      make_request(endpoint: 'ClaimantServiceBean/ClaimantWebService', action: 'findPOAByPtcpntId', body:,
+                   key: 'return')
+    end
   end
 end
